+++ ansitoimg/__main__.py
@@ -1,0 +1,4 @@
+""" entry point for python -m simplesecurity """
+from . import cli
+
+cli()