+++ foo.py
@@ -1,0 +1,5 @@
+from catimage.catimage import generateHDColour
+
+
+example3 = generateHDColour("test/test.png", 40)
+print(example3)