+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## 2020.4.1 - 2020/10/12
+- Add 'wide' parameter to renderers (enabled by default)
+  - true: width of 89, false: width of 49
+- Run from the command line with `ansitoimg`
+- set stdout to utf-8
+
 ## 2020.4 - 2020/10/11
 - Added typing (so dropping python < 3.7)
 - Support sgi chars in the form ESC[0Xm - fixing quite a lot of bugs